Subscription downloads: Add support for more than one subscription repository, also rename config section to something more neutral

+++ sitescripts/subscriptions/bin/updateEasyList.py
@@ -5,25 +5,35 @@
 # compliance with the License. You may obtain a copy of the License at
 # http://www.mozilla.org/MPL/
 
-import os, subprocess, tempfile, shutil
+import os, re, subprocess, tempfile, shutil
 from sitescripts.utils import get_config, setupStderr
 from sitescripts.subscriptions.combineSubscriptions import combineSubscriptions
 
 if __name__ == '__main__':
   setupStderr()
 
-  sourceRepo = get_config().get('easylist', 'repository')
-  destRepo = get_config().get('easylist', 'outdir')
+  sourceRepos = {}
+  for option, value in get_config().items('subscriptionDownloads'):
+    if option.endswith('_repository'):
+      sourceRepos[re.sub(r'_repository$', '', option)] = value
+  destDir = get_config().get('subscriptionDownloads', 'outdir')
 
-  sourceTemp = tempfile.mkdtemp()
-  destTemp = tempfile.mkdtemp()
+  sourceTemp = {}
+  destTemp = None
   try:
-    subprocess.Popen(['hg', 'archive', '-q', '-R', sourceRepo, '-r', 'default', sourceTemp]).communicate()
-    subprocess.Popen(['rsync', '-a', '--delete', destRepo + '/', destTemp]).communicate()
+    destTemp = tempfile.mkdtemp()
+    for repoName, repoDir in sourceRepos.iteritems():
+      tempDir = tempfile.mkdtemp()
+      sourceTemp[repoName] = tempDir
+      subprocess.Popen(['hg', 'archive', '-q', '-R', repoDir, '-r', 'default', tempDir]).communicate()
+#    subprocess.Popen(['rsync', '-a', '--delete', destDir + '/', destTemp]).communicate()
+    subprocess.Popen(['cp', '-r', destDir + '/.', destTemp]).communicate()
     combineSubscriptions(sourceTemp, destTemp)
-    subprocess.Popen(['rsync', '-au', '--delete', destTemp + '/', destRepo]).communicate()
+#    subprocess.Popen(['rsync', '-au', '--delete', destTemp + '/', destDir]).communicate()
+    subprocess.Popen(['cp', '-r', destTemp + '/.', destDir]).communicate()
   finally:
-    if os.path.exists(sourceTemp):
-      shutil.rmtree(sourceTemp, True)
-    if os.path.exists(destTemp):
+    for tempDir in sourceTemp.itervalues():
+      if os.path.exists(tempDir):
+        shutil.rmtree(tempDir, True)
+    if destTemp and os.path.exists(destTemp):
       shutil.rmtree(destTemp, True)
